fix(tests): complete MCP handshake helper, remove debug print, add initialization docs
