+++ tests/__init__.py
@@ -1,0 +1,3 @@
+# tests package
+__all__ = []
+
