+++ tests/test_lifespan.py
@@ -1,7 +1,6 @@
 import json
 import os
 import threading
-import time
 from unittest.mock import patch, AsyncMock
 
 import pytest
@@ -11,6 +10,8 @@ from httpx import AsyncClient
 
 from immich_mcp_server.immich_api import ImmichAPI
 from immich_mcp_server.server import mcp
+from tests.utils import initialize_session
+
 
 class UvicornTestServer(uvicorn.Server):
     """Uvicorn test server that runs in a thread."""
@@ -46,27 +47,7 @@ async def server():
     server.stop()
 
 
-async def initialize_session(client: AsyncClient) -> str:
-    """Helper to perform the MCP initialize handshake and return a session ID."""
-    init_request = {
-        "jsonrpc": "2.0",
-        "method": "initialize",
-        "params": {
-            "protocolVersion": "2025-06-18",
-            "clientInfo": {
-                "name": "immich-mcp-test-client",
-                "version": "0.1.0",
-            },
-            "capabilities": {},
-        },
-        "id": "init1",
-    }
-    headers = {"Accept": "application/json, text/event-stream"}
-    response = await client.post("/mcp", json=init_request, headers=headers)
-    response.raise_for_status()
-    # Give the server a moment to process the initialization
-    time.sleep(0.1)
-    return response.headers["mcp-session-id"]
+# `initialize_session` moved to `tests/utils.py` and imported above.
 
 
 @pytest.mark.asyncio
